requirements.txt
# Core
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1

# Database & ORM
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.12.0

# Forms & Validation
WTForms==3.0.1
email-validator==2.0.0

# Security
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0

# File handling
Pillow==10.0.0
python-magic==0.4.27

# Email & SMS
requests==2.31.0
africastalking==0.4.6

# Payment integrations
stripe==5.5.0
requests-mock==1.11.0  # For M-Pesa testing

# Utilities
python-dateutil==2.8.2
pandas==2.0.3
openpyxl==3.1.2

# Production
gunicorn==21.2.0
whitenoise==6.5.0

# Development (optional)
python-dotenv==1.0.0